Python script that scans for EC2 instances the scheduler instance has access to and stops/starts them accordingly.
Instances to be started must have the tag auto:start with a time defined in cron format. e.g. 10 13 * * 1-5 is 13:10 Monday to Friday.

Instances to be stopped must have the tag auto:stop with a time defined in cron format. e.g. 50 2 * * 2-6 is 02:50 Tuesday to Saturday.

Note also that time zone is in UTC so the 13:10 UTC startup will take place between 06:00 MST and 06:10 MST. While the 02:50 UTC stop will take place between 19:50 MST and 20:00 MST. This time range allows for the fact that the startup and shutdown operations may take several minutes.